# [WIP] Optimize geometry_union_agg and convex_hull_agg to reuse GeometryCursor