--- a/presto-geospatial/src/main/java/com/facebook/presto/plugin/geospatial/aggregation/GeometryCursorUtils.java
+++ b/presto-geospatial/src/main/java/com/facebook/presto/plugin/geospatial/aggregation/GeometryCursorUtils.java
@@ -1,0 +1,46 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.plugin.geospatial.aggregation;
+
+import com.esri.core.geometry.Geometry;
+import com.esri.core.geometry.GeometryCursor;
+import com.esri.core.geometry.ListeningGeometryCursor;
+import com.esri.core.geometry.ogc.OGCConcreteGeometryCollection;
+import com.esri.core.geometry.ogc.OGCGeometry;
+
+final class GeometryCursorUtils
+{
+    private GeometryCursorUtils() {}
+
+    static void processGeometry(OGCGeometry geometry, ListeningGeometryCursor input, GeometryCursor operator)
+    {
+        if (geometry == null || geometry.isEmpty()) {
+            return;
+        }
+        if (geometry instanceof OGCConcreteGeometryCollection) {
+            OGCConcreteGeometryCollection flattenedCollection = ((OGCConcreteGeometryCollection) geometry).flatten();
+            for (int i = 0; i < flattenedCollection.numGeometries(); i++) {
+                Geometry geometryN = flattenedCollection.geometryN(i).getEsriGeometry();
+                if (!geometryN.isEmpty()) {
+                    input.tick(geometryN);
+                    operator.tock();
+                }
+            }
+        }
+        else {
+            input.tick(geometry.getEsriGeometry());
+            operator.tock();
+        }
+    }
+}