--- a/presto-geospatial/src/main/java/com/facebook/presto/plugin/geospatial/aggregation/GeometryOperatorState.java
+++ b/presto-geospatial/src/main/java/com/facebook/presto/plugin/geospatial/aggregation/GeometryOperatorState.java
@@ -1,0 +1,39 @@
+/*
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package com.facebook.presto.plugin.geospatial.aggregation;
+
+import com.esri.core.geometry.ogc.OGCGeometry;
+import com.facebook.presto.spi.function.AccumulatorState;
+import com.facebook.presto.spi.function.AccumulatorStateMetadata;
+
+import static com.facebook.presto.plugin.geospatial.aggregation.GeometryOperatorStateFactory.GeometryConvexHullOperatorStateFactory;
+import static com.facebook.presto.plugin.geospatial.aggregation.GeometryOperatorStateFactory.GeometryUnionOperatorStateFactory;
+
+public interface GeometryOperatorState
+        extends AccumulatorState
+{
+    OGCGeometry getGeometry();
+
+    void add(OGCGeometry geometry);
+
+    @AccumulatorStateMetadata(stateSerializerClass = GeometryOperatorStateSerializer.class, stateFactoryClass = GeometryUnionOperatorStateFactory.class)
+    interface GeometryUnionOperatorState
+            extends GeometryOperatorState
+    {}
+
+    @AccumulatorStateMetadata(stateSerializerClass = GeometryOperatorStateSerializer.class, stateFactoryClass = GeometryConvexHullOperatorStateFactory.class)
+    interface GeometryConvexHullOperatorState
+            extends GeometryOperatorState
+    {}
+}