--- a/presto-geospatial/src/test/java/com/facebook/presto/plugin/geospatial/aggregation/TestGeometryUnionGeoAggregation.java
+++ b/presto-geospatial/src/test/java/com/facebook/presto/plugin/geospatial/aggregation/TestGeometryUnionGeoAggregation.java
@@ -271,18 +271,18 @@ public class TestGeometryUnionGeoAggregation
                                 "GEOMETRYCOLLECTION ( POLYGON (( 3 0, 3 2, 5 2, 5 0, 3 0 )), POLYGON (( 3 3, 3 5, 5 5, 5 3, 3 3 )) )"},
                 },
                 {
-                        "square with a line crossed becomes geometry collection",
-                        "GEOMETRYCOLLECTION (MULTILINESTRING ((0 2, 1 2), (3 2, 5 2)), POLYGON ((1 1, 3 1, 3 2, 3 3, 1 3, 1 2, 1 1)))",
+                        "square with a line crossed loses the line (lower dimension is dropped)",
+                        "POLYGON ((1 1, 3 1, 3 2, 3 3, 1 3, 1 2, 1 1))",
                         new String[] {"POLYGON ((1 1, 3 1, 3 3, 1 3, 1 1))", "LINESTRING (0 2, 5 2)"},
                 },
                 {
-                        "square with adjacent line becomes geometry collection",
-                        "GEOMETRYCOLLECTION (LINESTRING (0 5, 5 5), POLYGON ((1 1, 3 1, 3 3, 1 3, 1 1)))",
+                        "square with adjacent line loses the line (lower dimension is dropped)",
+                        "POLYGON ((1 1, 3 1, 3 3, 1 3, 1 1))",
                         new String[] {"POLYGON ((1 1, 3 1, 3 3, 1 3, 1 1))", "LINESTRING (0 5, 5 5)"},
                 },
                 {
-                        "square with adjacent point becomes geometry collection",
-                        "GEOMETRYCOLLECTION (POINT (5 2), POLYGON ((1 1, 3 1, 3 3, 1 3, 1 1)))",
+                        "square with adjacent point loses the point (lower dimension is dropped)",
+                        "POLYGON ((1 1, 3 1, 3 3, 1 3, 1 1))",
                         new String[] {"POLYGON ((1 1, 3 1, 3 3, 1 3, 1 1))", "POINT (5 2)"},
                 },
         };