Current geometry aggregation functions will serialize the input OGCGeometry and perform simple union or convexHull on them iteratively.  This works, but it suffers from poor memory locality, and it has a nontrivial object allocation overhead.  The idea of this PR is to use the internal operators that are invoked inside these two methods, to reuse the operators, and save on object allocation and get better in-memory performance.

The most significant effect of reusing the cursors is that we have access to the optimizations that exist inside them, instead of throwing them away after every iteration of the operation.  For example, for union, with this changeset, the individual union operations will now be grouped by coarsely defined bins which are based on the point count of the geometry with the number of dimensions, batched inside the union cursor, and once the internal batch size exceeds a constant value (10000), they will be unioned in aggregate.  This may result in an increase in memory, but the performance wins should outweigh the relative cost.  At the best case, where we continuously union geometries which overlap, this implementation will suffer from higher memory usage relative to the old implementation--for example, the points which resided inside other polygons will stick in memory longer.  At the worst case, where we are unioning non-overlapping geometries, there shouldn't be much difference in memory utilization, nor performance if the inputs are sufficiently large--once there are over 10000 non-overlapping geometries in a single node, performance may not be much better for whatever number that exceeds 10000.

I have added a JMH benchmark which performs union and convex hull on 50 states.  The performance improvements are stark.  Here are the results for the old method:

```
Result "com.facebook.presto.plugin.geospatial.BenchmarkGeometryAggregation.benchmarkConvexHull":
  23.716 ±(99.9%) 0.572 ms/op [Average]
  (min, avg, max) = (22.057, 23.716, 25.605), stdev = 0.856
  CI (99.9%): [23.144, 24.288] (assumes normal distribution)
```

```
Result "com.facebook.presto.plugin.geospatial.BenchmarkGeometryAggregation.benchmarkUnion":
  628.715 ±(99.9%) 10.353 ms/op [Average]
  (min, avg, max) = (601.299, 628.715, 671.803), stdev = 15.495
  CI (99.9%): [618.362, 639.068] (assumes normal distribution)
```

Here are the results for the new method:

```
Result "com.facebook.presto.plugin.geospatial.BenchmarkGeometryAggregation.benchmarkConvexHull":
  4.556 ±(99.9%) 0.182 ms/op [Average]
  (min, avg, max) = (4.124, 4.556, 5.187), stdev = 0.273
  CI (99.9%): [4.374, 4.739] (assumes normal distribution)
```

```
Result "com.facebook.presto.plugin.geospatial.BenchmarkGeometryAggregation.benchmarkUnion":
  21.780 ±(99.9%) 1.320 ms/op [Average]
  (min, avg, max) = (20.106, 21.780, 31.502), stdev = 1.976
  CI (99.9%): [20.460, 23.101] (assumes normal distribution)
```

There are a few things of note or which need to be fixed before this is ready for full review and merge:

1) There is no way presently to determine the memory utilization of a GeometryCursor.  This will need to be added to the Esri/geometry-api-java library.
2) In the proposed changes, there is a change in behavior for union, which will drop lower dimensions if two dimensional geometries are unioned.  For example, if one were to union a disjoint point and linestring, the union would return only the linestring.
3) There are some differences in behavior with convex hull which are either bugs in the library or something which needs to be fixed in this PR (some tests are failing for convex hull).
4) The above note regarding memory utilization needs to thoroughly studied.  If the 10000 number proves to be too large, we can easily modify the library to make it configurable.

I would be happy to receive feedback on this approach, given the caveats above.  CC: @mbasmanova